Moved registration to registration.py at the top level.
Fixed repr for base class NodeSet

Fixed documentation for build_blend_chain to reflect limitations of Maya blendColors node.
Fixed implementation for build_blend_chain as the zipping was not working properly based on the simplified inputs from the build fk/ik functions
Fixed the way Limb passed fk/ik chains to build_blend_chain.
Added Choice node to config.
