+++ anvil/config.py
@@ -105,6 +105,7 @@ IK_EFFECTOR = 'ik_effector'
 FK = 'fk'
 IKFK_BLEND = 'ikfk_blend'
 BLEND_NODE = 'blendColors'
+CHOICE_NODE = 'choice'
 POLE_VECTOR = 'pole_vector'
 BLEND = 'blend'
 WORLD_SPACE = 'worldSpace'
